reorganize prophet and tests
