+++ timeseries/src/autogluon/timeseries/models/gluonts/mx/models.py
@@ -11,8 +11,7 @@ from autogluon.timeseries.models.abstract.abstract_timeseries_model import Abstr
 from autogluon.timeseries.models.gluonts.abstract_gluonts import AbstractGluonTSModel
 
 with warning_filter():
-    from gluonts.model.estimator import Estimator as GluonTSEstimator, DummyEstimator
-    from gluonts.model.prophet import ProphetPredictor
+    from gluonts.model.estimator import Estimator as GluonTSEstimator
     from gluonts.mx.context import get_mxnet_context
     from gluonts.mx.model.deepar import DeepAREstimator
     from gluonts.mx.model.simple_feedforward import SimpleFeedForwardEstimator
@@ -460,54 +459,4 @@ class GenericGluonTSMXNetModelFactory(AbstractTimeSeriesModelFactory):
         return GenericGluonTSMXNetModel(
             gluonts_estimator_class=self.gluonts_estimator_class,
             **model_init_kwargs,
-        )
-
-
-class _ProphetDummyEstimator(DummyEstimator):
-    def train(self, train_data, validation_data=None, **kwargs):
-        return self.predictor
-
-
-class ProphetModel(AbstractGluonTSMXNetModel):
-    """Wrapper around `Prophet <https://github.com/facebook/prophet>`_, which wraps the
-    library through GluonTS's wrapper.
-
-    In order to use it you need to install the package::
-
-        pip install fbprophet
-
-    Other Parameters
-    ----------------
-    hyperparameters : Dict[str, Any]
-        Model hyperparameters that will be passed directly to the `Prophet`
-        class. See Prophet documentation for available parameters.
-    """
-
-    gluonts_estimator_class = _ProphetDummyEstimator
-    allowed_prophet_parameters = [
-        "growth",
-        "changepoints",
-        "n_changepoints",
-        "changepoint_range",
-        "yearly_seasonality",
-        "weekly_seasonality",
-        "daily_seasonality",
-        "holidays",
-        "seasonality_mode",
-        "seasonality_prior_scale",
-        "holidays_prior_scale",
-        "changepoint_prior_scale",
-        "mcmc_samples",
-        "interval_width",
-        "uncertainty_samples",
-    ]
-
-    def _get_estimator(self) -> GluonTSEstimator:
-        model_init_params = self._get_estimator_init_args()
-
-        return _ProphetDummyEstimator(
-            predictor_cls=ProphetPredictor,
-            freq=model_init_params["freq"],
-            prediction_length=model_init_params["prediction_length"],
-            prophet_params={k: v for k, v in model_init_params.items() if k in self.allowed_prophet_parameters},
         )
