+++ timeseries/tests/unittests/models/test_models.py
@@ -3,7 +3,6 @@ import itertools
 import os
 import shutil
 import tempfile
-from functools import partial
 from unittest import mock
 
 import numpy as np
@@ -17,7 +16,7 @@ from autogluon.timeseries.models.abstract import AbstractTimeSeriesModel
 
 from ..common import DUMMY_TS_DATAFRAME, dict_equal_primitive, get_data_frame_with_item_index
 from .test_autogluon_tabular import TESTABLE_MODELS as TABULAR_TESTABLE_MODELS
-from .test_gluonts import TESTABLE_MODELS as GLUONTS_TESTABLE_MODELS
+from .gluonts.test_gluonts import TESTABLE_MODELS as GLUONTS_TESTABLE_MODELS
 from .test_local import TESTABLE_MODELS as LOCAL_TESTABLE_MODELS
 from .test_sktime import TESTABLE_MODELS as SKTIME_TESTABLE_MODELS
 from .test_statsmodels import TESTABLE_MODELS as STATSMODELS_TESTABLE_MODELS
